feat(scraper): Add functions for scraping Prometheus metrics
Adds functions to discover scrapable Prometheus metrics from
Kubernetes. The discovery will be documented in the README and
does explicitly *not* support all discovery methods that Prometheus
itself has (yet).

Note that this is currently only functionining for local development
with `kubectl proxy` as the URLs are generated for that structure.
Code for in-cluster deployments exists but is not functional yet.